vix: extract primitive binding descriptors into runtime::primitive_bindings
Separate the compiler/binder's view of the builtin primitives (their
PrimitiveIds, request types, and surface-nameable wire types like OriginHint /
PinnedBlobRef / BlobId / ObserveRequest) from the RawPrimitive implementations
that execute them. The descriptors move into a new runtime/primitive_bindings.rs;
fetch_primitive / observe_primitive / tree_read_primitive keep only their impls
(now importing the descriptors from super::).

This pre-positions the language-side descriptors to stay in vix-core while the
impls move to vixen-primitives. Pure internal move; compiles clean.
