+++ vix/src/runtime/mod.rs
@@ -11,6 +11,7 @@ mod model;
 mod observe;
 mod observe_primitive;
 mod primitive;
+mod primitive_bindings;
 mod primitive_value_decode;
 mod scheduler;
 mod store;
@@ -28,6 +29,7 @@ pub use model::*;
 pub use observe::*;
 pub use observe_primitive::*;
 pub use primitive::*;
+pub use primitive_bindings::*;
 pub use primitive_value_decode::*;
 pub use scheduler::*;
 pub use store::*;
